requirements.txt
google-cloud-documentai==2.21.0
google-cloud-storage==2.10.0
Pillow==10.0.1
pandas==2.1.1
python-dotenv==1.0.0
streamlit==1.28.1
matplotlib==3.7.2
seaborn==0.12.2
reportlab==4.0.4
